Switch from x86_32 to x86_64.

+++ scripts/run.sh
@@ -14,23 +14,28 @@ set -e -u
 function print_usage {
 	printf "Usage: $0 [OPTIONS] <iso>\n"
 	printf "\n"
-	printf "\t-a <arch>             Architecture (Default: i386)\n"
 	printf "\t-d                    Debug mode\n"
 	printf "\t-g                    Freeze on boot and connect to GDB\n"
 	printf "\t-t                    Monitor mode\n"
 	printf "\t-k                    Enables kvm (if available)\n"
 	printf "\t-m <MB>               Memory (in MB) (Default: 512MB)\n"
-	printf "\t-c <CPU>              Cpu to emulate\n"
 	printf "\t-s <SMP>              Number of cores (Default: 1)\n"
 	printf "\t-h                    Print this help menu\n"
 	exit 1
 }
 
-# Qemu command and args
-declare qemu="qemu-system-i386"
-declare qemu_args=" -serial stdio -m 512"
+env > /tmp/jaj
 
-while getopts dgtkhm:a:c:s: FLAG; do
+# Check this script has ben launched the proper way.
+if [[ -z "${QEMU}" ]]; then
+	echo "Please use `make run` instead of using this script directly" > /dev/stder
+	exit 1
+fi
+
+declare qemu="$QEMU"
+declare qemu_args="$QEMU_ARGS ${QEMU_EXTRA_ARGS+}"
+
+while getopts dgtkhm:s: FLAG; do
 	case $FLAG in
 		d) qemu_args+=" -d int,cpu_reset,guest_errors,unimp --no-reboot";;
 		g) qemu_args+=" -s -S";;
@@ -38,12 +43,6 @@ while getopts dgtkhm:a:c:s: FLAG; do
 		t) qemu_args=$(echo "$qemu_args" | sed s/serial/monitor/g);;
 		m) qemu_args+=" -m $OPTARG";;
 		s) qemu_args+=" -smp $OPTARG";;
-		c) qemu_args+=" -cpu $OPTARG";;
-		a)
-			if [[ "$OPTARG" == "x86" ]]; then
-				OPTARG="i386"
-			fi
-			qemu="qemu-system-$OPTARG";;
 		h) print_usage;;
 		\?)
 			printf "Unknown option\n"
@@ -61,9 +60,9 @@ declare iso=$1
 qemu_args+=" -cdrom $iso"
 
 if ! which "$qemu" &> /dev/null; then
-	printf "  ERROR\t $qemu not found. You must install QEMU first.\n"
+	printf "  ERROR\t $qemu not found. You must install qemu first.\n"
 	exit 1
 fi
 
-printf "  QEMU\t $(basename $iso)\n"
+printf "  qemu\t $(basename $iso)\n"
 $qemu $qemu_args
